api/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.22.0
aioboto3>=11.0.0
python-multipart>=0.0.6
boto3>=1.28.0
botocore>=1.31.0
aiohttp>=3.8.5
motor>=3.3.0
pymongo>=4.6.0
aiofiles>=23.1.0
Pillow>=10.0.0
httpx>=0.27.0
requests